EnumIntegerField: Skip IntegerField validators
Make EnumIntegerField skip validators of django.db.models.IntegerField.

This is needed, because when database engine has limits for the integer
type (that is the case e.g. with MySQL), then the enum value will be
checked against those limits and that will result in following error:

  TypeError: unorderable types: TheEnum() < int()

Also add a test case for this.