Fix EnumFieldMixin.get_choices for enum34 1.1
On Python 2 with enum34 version 1.1.0 or newer, an enum item with zero
value will evaluate to false.  This makes `get_choices` return incorrect
data for the zero item.  Fix this by doing the checking if a choice is
an enum value with `if isinstance(i, Enum)` rather than with `if i`.

Add a test case that checks the choices are correct.