Add failing test that enum instances can be used as form data.
A common pattern in testing Django `ModelForm` behavior is to create a model instance,
then instantiate a form with modified data for that instance, then call `form.save()`,
and add assertions about what happened.

AFAICT, Django doesn't provide an officially supported way of doing this, but a
common method of building the data dictionary is to either:
1. Build the data dictionary by hand, or
2. Call the `model_to_dict` method and use that.

In the former case, you need to define the data dictionary like this:
```
data = {
		"my_field": MyEnum.THING.value,
		...
}
```
...which is annoying to remember.  In the latter case, you need to do a transform
to the data, finding all enums and accessing their `.value` attribute.